Move claim storage to insurance persistance

+++ contracts/Provident.sol
@@ -4,7 +4,7 @@ contract Provident {
   // Insurer related
   function getNumberOfInsurancePlans() constant public returns (uint16);
   function getInsurancePlanPrice(uint16 plan) constant public returns (uint256);
-  function getInsuredProfile(address insured) constant returns (int256 plan, int256 startDate, int256 finalDate);
+    function getInsuredProfile(address insured) constant returns (int16 plan, uint256 startDate, uint256 finalDate);
 
   function buyInsurancePlan(uint16 plan) payable public;
   function createClaim(uint16 claimType, string evidence, address beneficiary) returns (int);
